+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/pnl/ProfitLossPercentagePerTimeCriterionTest.java
@@ -1,0 +1,83 @@
+/**
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2014-2017 Marc de Verdelhan, 2017-2021 Ta4j Organization & respective
+ * authors (see AUTHORS)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package org.ta4j.core.analysis.criteria.pnl;
+
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import static org.ta4j.core.TestUtils.assertNumEquals;
+
+import java.time.temporal.ChronoUnit;
+import java.util.function.Function;
+
+import org.junit.Test;
+import org.ta4j.core.AnalysisCriterion;
+import org.ta4j.core.BaseTradingRecord;
+import org.ta4j.core.Position;
+import org.ta4j.core.Trade;
+import org.ta4j.core.TradingRecord;
+import org.ta4j.core.analysis.criteria.AbstractCriterionTest;
+import org.ta4j.core.mocks.MockBarSeries;
+import org.ta4j.core.num.Num;
+
+public class ProfitLossPercentagePerTimeCriterionTest extends AbstractCriterionTest {
+
+    public ProfitLossPercentagePerTimeCriterionTest(Function<Number, Num> numFunction) {
+        super((params) -> new ProfitLossPercentagePerTimeCriterion(ChronoUnit.SECONDS), numFunction);
+    }
+
+    @Test
+    public void calculateWithNoTrades() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 105, 110, 100, 95, 105);
+        assertNumEquals(0, getCriterion().calculate(series, new BaseTradingRecord()));
+    }
+
+    @Test
+    public void calculateWithTwoLongTrades() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 102, 104, 103, 102, 104);
+        TradingRecord tradingRecord = new BaseTradingRecord(Trade.buyAt(1, series), Trade.sellAt(5, series),
+                Trade.buyAt(3, series), Trade.sellAt(5, series));
+
+        assertNumEquals(0.586331620026651, getCriterion().calculate(series, tradingRecord));
+    }
+
+    @Test
+    public void calculateWithOneLongTrade() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 105, 110, 100, 95, 105);
+        Position position = new Position(Trade.buyAt(0, series), Trade.sellAt(4, series));
+
+        assertNumEquals(-1, getCriterion().calculate(series, position));
+    }
+
+    @Test
+    public void betterThan() {
+        AnalysisCriterion criterion = getCriterion();
+        assertTrue(criterion.betterThan(numOf(6), numOf(3)));
+        assertFalse(criterion.betterThan(numOf(4), numOf(7)));
+    }
+
+    @Test
+    public void testCalculateOneOpenPositionShouldReturnZero() {
+        openedPositionUtils.testCalculateOneOpenPositionShouldReturnExpectedValue(numFunction, getCriterion(), 0);
+    }
+}