+++ CHANGELOG.md
@@ -71,6 +71,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - :tada: **Enhancement** added **`GrossLossCriterion.class`**.
 - :tada: **Enhancement** added **`NetProfitCriterion.class`**.
 - :tada: **Enhancement** added chooseBest() method with parameter tradeType in AnalysisCriterion.
+- :tada: **Enhancement** added ProfitLossPercentagePerTimeCriterion.
 
 
 ## 0.13 (released November 5, 2019)